refactor: overhaul state resolution and related infrastructure for performance
This is a large refactor driven by performance issues in state_selector,
which lacked internal caching and became a bottleneck under stress tests.

Key changes include:
- Move Color from graphics to types
- Reorganize flor-macros dependency hierarchy to depend on flor-base and be used by flor
- Refactor state_selector into resolver, moving external trait-based computations
  into internal parameters with built-in caching and change detection
- Remove StateAccumulator as it is no longer necessary
- Consolidate all class parsing and extraction logic into shared.rs
- Change DPI type from f64 to f32
- Refactor UnitResolver
- Fix state computation being overwritten during scheduling
- Rename set_all_child_window_id to init_window_child and expand its responsibilities
- Remove bus_setup_arc_data, replaced by init_window_child and resolver logic
- Add control_state: ControlState parameter to on_measure
- Trigger layout update requests from builder reactive callbacks

+++ crates/platform/windows/src/window.rs
@@ -147,19 +147,19 @@ impl WindowApi for WindowId {
         todo!()
     }
 
-    fn get_dpi(&self) -> Result<(f64, f64), Self::Error> {
+    fn get_dpi(&self) -> Result<(f32, f32), Self::Error> {
         #[cfg(not(feature = "monitor"))]
         unsafe {
             #[cfg(feature = "win7-compat")]
             {
-                let dpi = crate::win7_compat::get_dpi_compatible(self.hwnd()) as f64;
+                let dpi = crate::win7_compat::get_dpi_compatible(self.hwnd()) as f32;
                 Ok((dpi, dpi))
             }
             #[cfg(not(feature = "win7-compat"))]
             {
                 use windows::Win32::UI::HiDpi::GetDpiForWindow;
                 let dpi = GetDpiForWindow(self.hwnd());
-                let val = if dpi == 0 { 96.0 } else { dpi as f64 };
+                let val = if dpi == 0 { 96.0 } else { dpi as f32 };
                 Ok((val, val))
             }
         }
